Introduce DatabaseClientManager vs DatabaseClient
