+++ libs/exo-sql/src/lib.rs
@@ -65,8 +65,9 @@ pub use sql::{
     array_util::{self, ArrayEntry},
     column::Column,
     database::{Database, TableId},
-    database_client::{
-        DatabaseClient, LOCAL_CHECK_CONNECTION_ON_STARTUP, LOCAL_CONNECTION_POOL_SIZE, LOCAL_URL,
+    database_client_manager::{
+        DatabaseClientManager, LOCAL_CHECK_CONNECTION_ON_STARTUP, LOCAL_CONNECTION_POOL_SIZE,
+        LOCAL_URL,
     },
     function::Function,
     limit::Limit,
