+++ libs/exo-sql/src/sql/mod.rs
@@ -20,6 +20,7 @@ pub mod array_util;
 pub mod column;
 pub mod database;
 pub mod database_client;
+pub mod database_client_manager;
 pub mod function;
 pub mod order;
 pub mod physical_column;
